requirements.txt
Flask>=2.3.3
Flask-SocketIO>=5.3.6
python-socketio>=5.10.0
python-engineio>=4.8.0
requests>=2.31.0
aiohttp>=3.9.1
pycoingecko>=3.1.0
pandas>=2.1.4
numpy>=1.26.2
web3>=6.11.3
python-dotenv>=1.0.0
solana>=0.30.2
websockets>=9.0,<12.0
python-binance>=1.0.19
eventlet>=0.33.3
gevent>=23.9.1
gevent-websocket>=0.10.1
PyQt5>=5.15.9
qasync>=0.27.1
